Second old code/Robot/robot_back/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
jproperties==2.1.2
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
waitress==3.0.0
